+++ rpc-framework/rpc-core/src/main/java/space/ruiwang/provider/RpcServer.java
@@ -1,8 +1,5 @@
 package space.ruiwang.provider;
 
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
 /**
  * RpcServer 接口，负责网络通信以及服务注册。
  * 可以通过 SPI 机制拓展不同的网络通信实现，目前已有 TomcatServer 的实现。
@@ -11,8 +8,5 @@ import java.util.concurrent.ConcurrentHashMap;
  * Created on 2025-02-11
  */
 public interface RpcServer {
-    // 使用线程安全的 Map 存储服务接口名与对应的服务实例
-    Map<String, Object> SERVICE_MAP = new ConcurrentHashMap<>();
-
     void start(String hostName, int port);
 }