+++ rpc-provider/src/main/java/space/ruiwang/serviceimpl/TestServiceImpl.java
@@ -7,7 +7,7 @@ import space.ruiwang.service.TestService;
  * @author wangrui <wangrui45@kuaishou.com>
  * Created on 2025-02-12
  */
-@RpcService()
+@RpcService
 public class TestServiceImpl implements TestService {
     @Override
     public String calc(Integer a, Integer b) {